backend/requirements.txt
numpy
pandas
scikit-learn==0.23.2
requests
torch
torchvision
Pillow